+++ pages/_app.tsx
@@ -2,15 +2,12 @@ import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 import RootLayout from './layout'
 import RootHead from './head'
-import styles from '../styles/Login.module.css'
-import { Button, Card, Checkbox, Form, Input, Space } from 'antd'
-import React from 'react'
-import { Typography } from "antd";
-
-const { Title } = Typography;
+import React, { useState } from 'react'
+import Login from '../components/login'
 
 export default function App({ Component, pageProps }: AppProps) {
-  const login: boolean = false;
+  const [login, setLogin] = useState(false)
+
   return (
     <>
       <RootHead />
@@ -20,71 +17,8 @@ export default function App({ Component, pageProps }: AppProps) {
           <Component {...pageProps} />
         </RootLayout>
       ) : (
-        <Login />
+        <Login login={login} setLogin={setLogin}  />
       )}
     </>
   );
 }
-
-const Login: React.FC = () => {
-  const onFinish = (values: any) => {
-    console.log("Success:", values);
-  };
-
-  const onFinishFailed = (errorInfo: any) => {
-    console.log("Failed:", errorInfo);
-  };
-  
-  return (
-    <main className={styles.login}>
-      <Space align="center" className={styles.loginPanel}>
-        <Card className={styles.loginCard}>
-          <Title>Log-In</Title>
-          <Form
-            name="basic"
-            labelCol={{ span: 8 }}
-            wrapperCol={{ span: 16 }}
-            initialValues={{ remember: true }}
-            onFinish={onFinish}
-            onFinishFailed={onFinishFailed}
-            autoComplete="off"
-          >
-            <Form.Item
-              label="Username"
-              name="username"
-              rules={[
-                { required: true, message: "Please input your username!" },
-              ]}
-            >
-              <Input />
-            </Form.Item>
-
-            <Form.Item
-              label="Password"
-              name="password"
-              rules={[
-                { required: true, message: "Please input your password!" },
-              ]}
-            >
-              <Input.Password />
-            </Form.Item>
-
-            <Form.Item
-              name="remember"
-              valuePropName="checked"
-              wrapperCol={{ offset: 8, span: 16 }}
-            >
-              <Checkbox>Remember me</Checkbox>
-            </Form.Item>
-
-            <Form.Item wrapperCol={{ offset: 8, span: 16 }}>
-              <Button type="primary" htmlType="submit">
-                Submit
-              </Button>
-            </Form.Item>
-          </Form>
-        </Card>
-      </Space>
-    </main>
-  );
-};
